arrange things sensibly
